Added simple client server UDP connection
Added a simple client server UDP connection. This is a simple call response connection where a client sends data to the server, the server response to the client, then the client sends one final message to the server before closing. This will need worked into our bevy framework but serves as a decent intro to a small UDP back and forth. Connection is currently on localhost 127.0.0.1 with the server on port 4000, and the client on port 8000.